CHORE : move types directory
