+++ frontend/src/types/Friends.tsx
