+++ frontend/src/types/MyInfo.tsx
