+++ frontend/src/types/UserDetail.tsx
